requirements.txt
pandas
pyyaml
python-dotenv
